requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Forecasting (optional - for ARIMA)
statsmodels==0.14.1

# File uploads
python-multipart==0.0.6

# Date handling
python-dateutil==2.8.2
